tambah throw not found

+++ src/api/cartdetails/cartdetails.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCartdetailDto } from './dto/create-cartdetail.dto';
 import { UpdateCartdetailDto } from './dto/update-cartdetail.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -20,14 +20,20 @@ export class CartdetailsService {
   }
 
   async findOne(id: number) {
-    return await this.cartdetailrepo.findOne({where: {id:id}, relations:{cart:true, product: true}});
+    var data = await this.cartdetailrepo.findOne({where: {id:id}, relations:{cart:true, product: true}});
+    if(!data) throw new NotFoundException();
+    return data;
   }
 
   async update(id: number, updateCartdetailDto: UpdateCartdetailDto) {
+    var data = await this.cartdetailrepo.findOne({where: {id:id}});
+    if(!data) throw new NotFoundException();
     return await this.cartdetailrepo.update({id}, {...updateCartdetailDto, updatedAt: new Date()});
   }
 
   async remove(id: number) {
+    var data = await this.cartdetailrepo.findOne({where: {id:id}});
+    if(!data) throw new NotFoundException();
     return await this.cartdetailrepo.delete({id});
   }
 }
