+++ src/api/orders/orders.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { UpdateOrderDto } from './dto/update-order.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -13,9 +13,6 @@ export class OrdersService {
   async create(createOrderDto: CreateOrderDto) {
     const order = this.orderrepo.create(createOrderDto);
     await this.orderrepo.save(order);
-    // if(order.orderdetails.length > 0){
-    //   return order.orderdetails.length;
-    // }
   }
 
   async findAll() {
@@ -23,14 +20,20 @@ export class OrdersService {
   }
 
   async findOne(id: number) {
-    return await this.orderrepo.findOne({where: {id:id}, relations:{user:true, delivery: true, payment: true, orderdetails: true}});
+    var data = await this.orderrepo.findOne({where: {id:id}, relations:{user:true, delivery: true, payment: true, orderdetails: true}});
+    if(!data) throw new NotFoundException();
+    return data;
   }
 
   async update(id: number, updateOrderDto: UpdateOrderDto) {
+    var data = await this.orderrepo.findOne({where: {id:id}});
+    if(!data) throw new NotFoundException();
     return await this.orderrepo.update({id}, {...updateOrderDto, updatedAt: new Date()});
   }
 
   async remove(id: number) {
+    var data = await this.orderrepo.findOne({where: {id:id}});
+    if(!data) throw new NotFoundException();
     return await this.orderrepo.delete({id});
   }
 }
